feat: 점수 계산 기능 분리
